+++ clinica/pipelines/dwi_connectome/dwi_connectome_utils.py
@@ -15,33 +15,59 @@ def get_luts():
     return [default, a2009s]
 
 
+def get_conversion_luts_offline():
+
+    # TODO: use this function if no internet connect found in client (need to upload files to clinica repository)
+    return
+
+
 def get_conversion_luts():
-    import os
+    from os import pardir
+    from os.path import abspath, dirname, join
+    from pathlib import Path
 
-    from clinica.utils.exceptions import ClinicaException
+    from clinica.utils.inputs import RemoteFileStructure, fetch_file
+    from clinica.utils.stream import cprint
 
-    try:
-        # For aparc+aseg.mgz file:
-        default = os.path.join(
-            os.environ["MRTRIX_HOME"],
-            "share",
-            "mrtrix3",
-            "labelconvert",
-            "fs_default.txt",
-        )
-        # For aparc.a2009s+aseg.mgz file:
-        a2009s = os.path.join(
-            os.environ["MRTRIX_HOME"],
-            "share",
-            "mrtrix3",
-            "labelconvert",
-            "fs_a2009s.txt",
-        )
+    root = dirname(abspath(join(abspath(__file__), pardir, pardir)))
 
-        # TODO: Add custom Lausanne2008 conversion LUTs here.
-    except KeyError:
-        raise ClinicaException("Could not find MRTRIX_HOME environment variable.")
-    return [default, a2009s]
+    path_to_mappings = Path(root) / "resources" / "mappings"
+
+    url_mrtrix = "https://raw.githubusercontent.com/MRtrix3/mrtrix3/master/share/mrtrix3/labelconvert/"
+
+    fs_default = RemoteFileStructure(
+        filename="fs_default.txt",
+        url=url_mrtrix,
+        checksum="6ee07088915fdbcf52b05147ddae86e5fcaf3efc63db5b0ba8f361637dfa11ef",
+    )
+
+    fs_a2009s = RemoteFileStructure(
+        filename="fs_a2009s.txt",
+        url=url_mrtrix,
+        checksum="b472f09cfe92ac0b6694fb6b00a87baf15dd269566e4a92b8a151ff1080bf170",
+    )
+
+    ref_fs_default = path_to_mappings / Path(fs_default.filename)
+    ref_fs_a2009 = path_to_mappings / Path(fs_a2009s.filename)
+
+    if not (ref_fs_default.is_file()):
+        try:
+            ref_fs_default = fetch_file(fs_default, path_to_mappings)
+        except IOError as err:
+            cprint(
+                msg=f"Unable to download required MRTRIX mapping (fs_default.txt) for processing: {err}",
+                lvl="error",
+            )
+    if not (ref_fs_a2009.is_file()):
+        try:
+            ref_fs_a2009 = fetch_file(fs_a2009s, path_to_mappings)
+        except IOError as err:
+            cprint(
+                msg=f"Unable to download required MRTRIX mapping (fs_a2009s.txt) for processing: {err}",
+                lvl="error",
+            )
+
+    return [ref_fs_default, ref_fs_a2009]
 
 
 def get_containers(subjects, sessions):