+++ docs/CHANGELOG.md
@@ -6,6 +6,20 @@ Main changes to this code/ project are documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Clinica 0.7.1
+
+
+### Added
+- [Doc] add ukbiobank documentation
+- [DWIConnectome] Fetch meta data directly from MRtrix github repository
+
+### Changed
+- [Core] Enable parallelization when grabbing files
+
+### Fixed
+- [Converters] Fix several warnings
+
+
 ## Clinica 0.7.0
 
 ### Added